Rename library filename
