+++ src/index.ts
@@ -1,1 +1,1 @@
-export { BskyTldr } from './lib/bluesky-generators';
+export { BskyTldr } from './lib/bsky-tldr';
